# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "prempawar130@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/downloadusr": {
            "get": {
                "tags": [
                    "prem"
                ],
                "summary": "List all studets",
                "description": "create folder<br/>\n       Success Code:<br/>\n       8000: Response-Note updated successfully.<br/>\n       Error Code:<br/>\n       10001: Invalid session-token or its expired.<br/>\n       7031: Please enter a valid subscription(1/2/3).<br/>\n       7032: Please enter a valid sort by.<br/>\n       7033: Please enter a valid order(ASC/DESC).<br/>\n       ",
                "operationId": "App\\Http\\Controllers\\downloadUsrController::export",
                "parameters": [
                    {
                        "name": "session-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_By",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "con_first_name/con_last_name"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "ASC/DESC"
                        }
                    },
                    {
                        "name": "subscription",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "0/1"
                        }
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Please enter a valid information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/swa_parameter": {
            "get": {
                "tags": [
                    "PREM Tag"
                ],
                "summary": "Get list of available subscription plans for user",
                "description": "Subscription list",
                "operationId": "App\\Http\\Controllers\\parametercontroller::parameter",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}